+++ index.html
@@ -8,7 +8,9 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
+  <script src="/js/data.js" type="module"></script>
+  <script src="/js/util.js" type="module"></script>
   <title>Кекстаграм</title>
 </head>
